Pull request #5181: payload_injector: add support for payload injection on s2c packets for http and http2 traffic
Merge in SNORT/snort3 from ~ADMAMOLE/snort3:test_s2c to master

Squashed commit of the following:

commit 5f0f358b3c2864c2a11d9697c8ce5046c2dfa7b6
Author: Adrian Mamolea <admamole@cisco.com>
Date:   Thu Feb 26 13:58:35 2026 -0500

    payload_injector: add support for payload injection on s2c packets for http and http2 traffic

+++ src/stream/tcp/tcp_segment_descriptor.h
@@ -128,10 +128,10 @@ public:
         pkt->dsize -= offset;
     }
 
-    void set_packet_flags(uint32_t flags) const
+    void set_packet_flags(uint64_t flags) const
     { pkt->packet_flags |= flags; }
 
-    bool are_packet_flags_set(uint32_t flags) const
+    bool are_packet_flags_set(uint64_t flags) const
     { return (pkt->packet_flags & flags) == flags; }
 
     uint32_t get_packet_timestamp() const
